cmv_patients/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
python-dotenv
alembic
cryptography
python-jose
passlib
bcrypt
httpx
fastapi_limiter
email-validator